chore: ios support on components (#56)
* chore: ios support on components

* chore: adjustments

* chore: adding missing rule

* chore: adding missing rule

* chore: disabling minifying

+++ components/src/commonMain/kotlin/me/lincolnstuart/funblocks/essentials/chart/bar/utils/BarChartValue.kt
@@ -1,15 +1,14 @@
 package me.lincolnstuart.funblocks.essentials.chart.bar.utils
 
 import kotlinx.datetime.LocalDate
-import java.math.BigDecimal
 
 /**
  * Data class that a represents the cartesian point that the bar have to grow up until.
  *
- * @param value [BigDecimal] y coordinate.
+ * @param value [Double] y coordinate.
  * @param date [LocalDate] x coordinate.
  */
 public class BarChartValue(
-    val value: BigDecimal,
+    val value: Double,
     val date: LocalDate
 )
